Move CU modal icon
